update as of 9:33pm, 16 April 2017
I have everything set up so that it should be working, but it isn't. The client sends to the router, the router says it connects to the other router but the other router does not accept the connection on the listen socket. I have no idea why, the router is able to receive messages from the client running on the local ip of 127.0.0.1 but fails to receive when using an external ip such as 157.160.37.131. The only change is that it is coming from an external ip. My only thinking is that the operating system is blocking the request for some reason..........